+++ deepinv/physics/generator/mri.py
@@ -81,6 +81,14 @@ class BaseMaskGenerator(PhysicsGenerator, ABC):
         """
         pass
 
+    @abstractmethod
+    def get_pdf(self) -> torch.Tensor:
+        """Get mask PDF.
+
+        :return torch.Tensor: unnormalised 1D vector representing pdf evaluated across mask columns.
+        """
+        pass
+
     def step(
         self, batch_size=1, seed: int = None, img_size: Optional[Tuple] = None, **kwargs
     ) -> dict:
@@ -175,6 +183,88 @@ class RandomMaskGenerator(BaseMaskGenerator):
         return mask
 
 
+class PolyOrderMaskGenerator(BaseMaskGenerator):
+    """
+    Generator for MRI Cartesian acceleration masks using polynomial variable density.
+
+    Generates a mask of vertical lines for MRI acceleration with fixed sampling in low frequencies (center of k-space) and polynomial order sampling in the high frequencies.
+    Polynomial sampling varies in 1D as :math:`(1 - r)^{p}` where :math:`r` is the distance from the centre and :math:`p` is the polynomial order.
+
+    This is achieved by creating a 1D polynomial variable density function then using this to sample from a Bernoulli.
+
+    The mask is repeated across channels and randomly varies across batch dimension.
+
+    Algorithm taken from `Millard and Chiew <https://pmc.ncbi.nlm.nih.gov/articles/PMC7614963/>`_.
+
+    :Examples:
+
+        >>> from deepinv.physics.generator.mri import PolyOrderMaskGenerator
+        >>> generator = PolyOrderMaskGenerator((2, 128, 128), acceleration=8, center_fraction=0.04, poly_order=8)
+        >>> params = generator.step(batch_size=1)
+        >>> mask = params["mask"]
+        >>> mask.shape
+        torch.Size([1, 2, 128, 128])
+
+    For other parameter descriptions see :class:`deepinv.physics.generator.mri.BaseMaskGenerator`
+
+    :param int poly_order: polynomial order of the sampling pdf
+    """
+
+    def __init__(self, *args, poly_order: int = 8, **kwargs):
+        super().__init__(*args, **kwargs)
+        self.poly_order = poly_order
+        self.pdf = self.get_pdf()
+
+    def get_pdf(self, max_iter: int = 100, tol: float = 1e-3) -> torch.Tensor:
+        # Distance from the center (normalized to [-1, 1])
+        r = torch.linspace(-1, 1, self.W, device=self.device).abs()
+
+        # Polynomial variable density function
+        pdf = (1 - r) ** self.poly_order
+
+        # Sample the central lines
+        pdf[
+            self.W // 2 - self.n_center // 2 : self.W // 2 + ceildiv(self.n_center, 2)
+        ] = 1
+
+        # Using binary search to scale the mask to the desired acceleration factor
+        a, b = -1.0, 1.0  # Binary search bounds
+        target_fraction = 1.0 / self.acc
+
+        for i in range(max_iter):
+            c = (a + b) / 2  # Midpoint of search range
+            scaled_pdf = pdf + c  # Adjust probabilities
+            scaled_pdf.clamp_(0, 1)  # Keep values in range [0, 1]
+            scaled_pdf[
+                self.W // 2
+                - self.n_center // 2 : self.W // 2
+                + ceildiv(self.n_center, 2)
+            ] = 1  # Ensure center is fully sampled
+
+            # Calculate the current sampling fraction
+            current_fraction = scaled_pdf.mean().item()
+
+            # Adjust binary search bounds
+            if current_fraction < target_fraction - tol:
+                a = c
+            elif current_fraction > target_fraction + tol:
+                b = c
+            else:
+                return scaled_pdf
+        else:
+            raise ValueError(f"get_pdf did not converge after {max_iter} iterations")
+
+    def sample_mask(self, mask: torch.Tensor) -> torch.Tensor:
+        for b in range(mask.shape[0]):
+            for t in range(mask.shape[2]):
+                mask[b, :, t, :, :] = (
+                    torch.bernoulli(self.pdf, generator=self.rng)
+                    .unsqueeze(0)
+                    .expand(self.H, self.W)
+                )
+        return mask
+
+
 class GaussianMaskGenerator(RandomMaskGenerator):
     """Generator for MRI Cartesian acceleration masks using Gaussian undersampling.
 
@@ -245,6 +335,9 @@ class EquispacedMaskGenerator(BaseMaskGenerator):
         torch.Size([1, 2, 8, 64, 64])
 
     """
+
+    def get_pdf(self) -> torch.Tensor:
+        raise NotImplementedError("get_pdf is undefined for this mask generator.")
 
     def sample_mask(self, mask: torch.Tensor) -> torch.Tensor:
         _W = mask.shape[-1]